+++ LightBlue.Tests/LightBlue.Tests.csproj
@@ -99,7 +99,7 @@
     <Compile Include="Standalone\StandaloneAzureLocalResourceSourceTests.cs" />
     <Compile Include="Standalone\StandaloneAzureSettingsTests.cs" />
     <Compile Include="Standalone\StandaloneAzureTestsBase.cs" />
-    <Compile Include="Standalone\WaitExtensionsTests.cs" />
+    <Compile Include="Standalone\FileLockExtensionsTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\TestCloudService\ServiceConfiguration.Local.cscfg">
